requirements.txt
# dbt-core==1.9.0
# dbt-postgres==1.9.0
# # dbt-bigquery==1.9.0
# elementary-data==0.16.1
# psycopg2-binary==2.9.10
# astronomer-cosmos==1.7.1
# apache-airflow==2.10.4
# celery==5.4.0
# protobuf==5.29.1
# sqlfluff
# sqlfluff-templater-dbt
# ruff
# pre-commit
# pytest
# fabric
psycopg2-binary
dbt-core
dbt-postgres
# dbt-bigquery==1.9.0
elementary-data
astronomer-cosmos
apache-airflow
celery
protobuf